refactor: rename druid_controller module to controller
Frees up the `controller/` directory for upcoming dereference and validate
submodules, matching the trino-operator layout.

No behavior change.
